Add files via upload
After researching about the Foursquare API I have used some of their suggested code to understand how it works, initially by creating a text document to see how the data is represented. I then created my for loop to get the name of every place that is nearby in a 500 metre radius.